Fixed bugs with the 0x Exchange interface. Imported 0x Exchange and TokenTransferProxy into project for redeployment to the testnet.

+++ migrations/2_deploy_tokens.js
@@ -8,7 +8,7 @@ var fs = require('fs');
 
 module.exports = function(deployer, network, accounts) {
 
-	console.log("before balance: "+web3.eth.getBalance(accounts[0]));
+	console.log("migrations :: before balance: "+web3.eth.getBalance(accounts[0]));
 
 	deployer.deploy(b0xToken).then(function() {
 		b0xToken.deployed().then(function(instance) {
@@ -19,33 +19,35 @@ module.exports = function(deployer, network, accounts) {
 
 
 	for (var i = 0; i < 10; ++i) {
-		deployer.new(
-			BaseToken,
-			10000000000000000000000000,
-			"TestToken"+i, 
-			18, 
-			"TEST"+i).then(function(instance) {
-				return instance.name.call().then(function(name) {
-					console.log(name + " created: "+instance.address);
-					fs.writeFile("./build/contracts/"+name+".json", 
-						JSON.stringify(
-						{
-							"contractName": name,
-							"abi": instance.abi,
-							"networks": {
-								"50": {
-									"events": {},
-									"links": {},
-									"address": instance.address
+		if (!fs.existsSync("./build/contracts/TestToken"+i+".json")) {
+			deployer.new(
+				BaseToken,
+				10000000000000000000000000,
+				"TestToken"+i, 
+				18, 
+				"TEST"+i).then(function(instance) {
+					return instance.name.call().then(function(name) {
+						//console.log(name + " created: "+instance.address);
+						fs.writeFile("./build/contracts/"+name+".json", 
+							JSON.stringify(
+							{
+								"contractName": name,
+								"abi": instance.abi,
+								"networks": {
+									"50": {
+										"events": {},
+										"links": {},
+										"address": instance.address
+									}
 								}
+							}), function(err) {
+							if(err) {
+								return console.log(err);
 							}
-						}), function(err) {
-						if(err) {
-							return console.log(err);
-						}
+						});
 					});
-				});
-		});
+			});
+		}
 	}
 
 	/*deployer.deploy(TOMToken).then(function() {
